+++ esphome/components/zyaura/zyaura.cpp
@@ -6,7 +6,7 @@ namespace zyaura {
 
 static const char *TAG = "zyaura";
 
-bool ICACHE_RAM_ATTR ZaDataProcessor::decode(unsigned long ms, bool data) {
+bool ICACHE_RAM_ATTR ZaDataProcessor::decode(uint32_t ms, bool data) {
   // check if a new message has started, based on time since previous bit
   if ((ms - this->prev_ms_) > ZA_MAX_MS) {
     this->num_bits_ = 0;
